Features implementation. Workflow Settings Plugin modified

+++ src/designer/elsa-workflows-studio/src/modules/elsa-workflows-settings/plugins/workflow-settings-feature-plugin.tsx
@@ -1,7 +1,7 @@
 import {Component, Prop, State, h} from '@stencil/core';
 import {createElsaWorkflowSettingsClient, SaveWorkflowSettingsRequest} from "../services/elsa-client";
 import {eventBus} from "../../../services";
-import {EventTypes, ConfigureFeatureContext, WorkflowSettingsUpdatedContext} from "../../../models";
+import {EventTypes, ConfigureFeatureContext} from "../../../models";
 import {WorkflowSettings} from "../models";
 
 @Component({
@@ -14,37 +14,39 @@ export class ElsaWorkflowSettingsFeaturePlugin {
 
   connectedCallback() {
     eventBus.on(EventTypes.ConfigureFeature, this.onWorkflowSettingsEnabled);
-    //eventBus.on(EventTypes.WorkflowSettingsUpdated, this.onWorkflowSettingsUpdated);
+    eventBus.on(EventTypes.FeatureUpdated, this.onWorkflowSettingsUpdated);
   }
 
   disconnectedCallback() {
     eventBus.detach(EventTypes.ConfigureFeature, this.onWorkflowSettingsEnabled);
-    //eventBus.detach(EventTypes.WorkflowSettingsUpdated, this.onWorkflowSettingsUpdated);
+    eventBus.detach(EventTypes.FeatureUpdated, this.onWorkflowSettingsUpdated);
   }
   
   onWorkflowSettingsEnabled(context: ConfigureFeatureContext) {
     if (context.featureName != "settings")
       return;
 
-    context.isEnabled = true;
     context.headers.push({url: null, label: "Enabled", component: null, exact: false})
     context.columns.push({url: null, label: "Enabled", component: null, exact: false})
     context.hasContextItems = true;
   }  
 
-  async onWorkflowSettingsUpdated(context: WorkflowSettingsUpdatedContext) {
+  async onWorkflowSettingsUpdated(context: string) {
+    let data: string[] = JSON.parse(context);    
+    if (data[0] != "settings")
+      return;
 
     const elsaClient = createElsaWorkflowSettingsClient(this.serverUrl);    
     this.workflowSettings = await elsaClient.workflowSettingsApi.list();
 
-    const workflowBlueprintSettings = this.workflowSettings.find(x => x.workflowBlueprintId == context.workflowBlueprintId && x.key == context.key);
+    const workflowBlueprintSettings = this.workflowSettings.find(x => x.workflowBlueprintId == data[1] && x.key == data[2]);
     if (workflowBlueprintSettings != undefined)
       await elsaClient.workflowSettingsApi.delete(workflowBlueprintSettings.id);
-
+    
     const request: SaveWorkflowSettingsRequest = {
-      workflowBlueprintId : context.workflowBlueprintId,
-      key: context.key,
-      value: context.value
+      workflowBlueprintId : data[1],
+      key: data[2],
+      value: data[3]
     };
 
     await elsaClient.workflowSettingsApi.save(request);
